Move data preview types to shared interface
